# northbound: KISS always batch yang config, it's faster.