The backoff code assumed that yang operations always completed quickly.
It checked for > 100 YANG modeled commands happening in under 1 second
to enable batching. If 100 yang modeled commands always take longer than
1 second batching is never enabled. This is the exact opposite of what
we want to happen since batching speeds the operations up.

Example:

Old code: adding 1000 static routes (which causes future route adds to
take longer) then adding another 1000 static routes: the first 1000 add
in ~3s, the second 1000 routes add in ~59 seconds.

New code: adding 1000 static routes takes ~3s, and adding a
second 1000 takes ~9s.

Future benefit, with libyang2 the old code adds are ~1s and ~115s, and
the new code adds are ~1s and ~2s. An even better improvement.