@rzalamena 2021-03-07
Similar PR here: https://github.com/FRRouting/frr/pull/7786

You can fix the `make check` (unit tests) with this commit:
https://github.com/FRRouting/frr/pull/7786/commits/a6478178e71e96b587a3e583eaed55c22dca2ae6

However it won't fix all other topotest that rely on instantaneous command application.

---

About this PR specifically: did you limit the amount of commands to batch? I think it might not be a good idea to batch all possible commands or it might blow things up on huge configs.