@rwestphal 2021-03-14
> The backoff code assumed that yang operations always completed quickly.
> It checked for > 100 YANG modeled commands happening in under 1 second
> to enable batching. If 100 yang modeled commands always take longer than
> 1 second batching is never enabled. This is the exact opposite of what
> we want to happen since batching speeds the operations up.

Yeah, the backoff logic is indeed flawed as it doesn't take into account the time spent processing the commands themselves.

My only concern with these changes is this thing I mentioned in the original backoff PR (#6727):
> The downside of this optimization is that, if several YANG-modeled commands are grouped into the same transaction and at least one of them fails, the whole transaction is rejected. This is undesirable since users don't expect transactional behavior when that's not explicitly enabled.

So the more we use command batching, the higher the chances of that problem occurring. It might be a fair price to pay to solve our NB performance problems, at least for the time being, but it needs to be a conscious decision taking that into account.