@choppsv1 2021-05-31
> And I have one more question:
> Previously the batching was working not only when reading the config but also when the user entered more than 100 commands in a short period of time. Now it works only during startup. Did I get this right?

That is correct. This eliminates the use of a timer for applying commits which had issues.