@choppsv1 2021-05-31
I made one small change outside of what has been reviewed, I changed `lyd_free_tree` to `lyd_free_all` of the temporary diff that is created and freed in `nb_config_diff`. This was a memory leak identified by valgrind, and it's easier to fix it here in this to be back-ported PR than to do another entire PR for it.